Completed filter functionality by assigning CSS classes instead of tampering with the array
