+++ src/assets/main.css
@@ -85,6 +85,14 @@ button {
     font-family: rockwell;
 }
 
+.incompleteTask {
+    width: 100%;
+    height: 100%;
+    font-size: 30px;
+    color: black;
+    font-family: rockwell;
+}
+
 .removeTaskButton {
     border-radius: 5px;
     height: 100%;
@@ -93,7 +101,6 @@ button {
     text-decoration: revert;
     all: revert;
 }
-
 
 .taskContent {
     display: flex;
@@ -105,6 +112,7 @@ button {
     flex-flow: nowrap;
     justify-content: center;
     margin-top: 50px;
+    margin-bottom: 100px;
 }
 
 .filterList {
@@ -113,22 +121,26 @@ button {
     justify-content: space-between;
 }
 
-.allTasksCheckbox {
+.allTasksRadioButton {
     margin-right: 3em;
     font-size: 25px;
     font-family: rockwell;
     color: black;
 }
 
-.completeTasksCheckbox {
+.completeTasksRadioButton {
     margin-right: 3em;
     font-size: 25px;
     font-family: rockwell;
     color: black;
 }
 
-.incompleteTasksCheckbox {
+.incompleteTasksRadioButton {
     font-size: 25px;
     font-family: rockwell;
     color: black;
+}
+
+.hideTask {
+    display: none;
 }
